+++ spreadsheet-api/src/main/java/ec/util/spreadsheet/Book.java
@@ -1,17 +1,17 @@
 /*
  * Copyright 2013 National Bank of Belgium
  *
- * Licensed under the EUPL, Version 1.1 or – as soon they will be approved 
+ * Licensed under the EUPL, Version 1.1 or – as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
  *
  * http://ec.europa.eu/idabc/eupl
  *
- * Unless required by applicable law or agreed to in writing, software 
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
 package ec.util.spreadsheet;
@@ -30,8 +30,7 @@ import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
-import java.util.Date;
-import java.util.Objects;
+import java.util.*;
 import java.util.function.ObjIntConsumer;
 import nbbrd.service.Mutability;
 import nbbrd.service.Quantifier;
@@ -281,6 +280,10 @@ public abstract class Book implements Closeable {
          */
         abstract public void store(@NonNull OutputStream stream, @NonNull Book book) throws IOException;
         //</editor-fold>
+
+        public @NonNull Map<String, List<String>> getExtensionsByMediaType() {
+            return Collections.emptyMap();
+        }
 
         @Override
         public boolean accept(Path entry) throws IOException {